Homework03/requirements.txt
boto3==1.12.5
botocore==1.15.5
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
docutils==0.15.2
ez-setup==0.9
Flask==1.1.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
gae-sdk-path-fixer==0.1.2
gcloud==0.18.3
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-logging==1.15.0
google-cloud-datastore==1.11.0
google-cloud-storage==1.26.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
httplib2==0.17.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
MarkupSafe==1.1.1
mongoengine==0.19.1
oauth2client==4.1.3
oauthlib==3.1.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.23.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.3.3
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.0
WTForms==2.2.1
